Move stage 1 and 2 files to their own directories
