+++ src/arm64/stage1_find_marks.h
@@ -71,17 +71,17 @@ really_inline simd8<bool> must_be_continuation(simd8<uint8_t> prev1, simd8<uint8
     return is_second_byte ^ is_third_byte ^ is_fourth_byte;
 }
 
-#include "generic/buf_block_reader.h"
-#include "generic/json_string_scanner.h"
-#include "generic/json_scanner.h"
+#include "generic/stage1/buf_block_reader.h"
+#include "generic/stage1/json_string_scanner.h"
+#include "generic/stage1/json_scanner.h"
 
-#include "generic/json_minifier.h"
+#include "generic/stage1/json_minifier.h"
 WARN_UNUSED error_code implementation::minify(const uint8_t *buf, size_t len, uint8_t *dst, size_t &dst_len) const noexcept {
   return arm64::stage1::json_minifier::minify<64>(buf, len, dst, dst_len);
 }
 
-#include "generic/utf8_lookup2_algorithm.h"
-#include "generic/json_structural_indexer.h"
+#include "generic/stage1/utf8_lookup2_algorithm.h"
+#include "generic/stage1/json_structural_indexer.h"
 WARN_UNUSED error_code implementation::stage1(const uint8_t *buf, size_t len, parser &parser, bool streaming) const noexcept {
   return arm64::stage1::json_structural_indexer::index<64>(buf, len, parser, streaming);
 }
