+++ src/arm64/stage2_build_tape.h
@@ -9,10 +9,10 @@
 namespace simdjson {
 namespace arm64 {
 
-#include "generic/atomparsing.h"
-#include "generic/structural_iterator.h"
-#include "generic/structural_parser.h"
-#include "generic/streaming_structural_parser.h"
+#include "generic/stage2/atomparsing.h"
+#include "generic/stage2/structural_iterator.h"
+#include "generic/stage2/structural_parser.h"
+#include "generic/stage2/streaming_structural_parser.h"
 
 } // namespace arm64
 } // namespace simdjson
