+++ src/arm64/stringparsing.h
@@ -45,7 +45,7 @@ really_inline backslash_and_quote backslash_and_quote::copy_and_find(const uint8
   };
 }
 
-#include "generic/stringparsing.h"
+#include "generic/stage2/stringparsing.h"
 
 } // namespace arm64
 } // namespace simdjson
