+++ src/fallback/stringparsing.h
@@ -27,7 +27,7 @@ really_inline backslash_and_quote backslash_and_quote::copy_and_find(const uint8
   return { src[0] };
 }
 
-#include "generic/stringparsing.h"
+#include "generic/stage2/stringparsing.h"
 
 } // namespace fallback
 } // namespace simdjson
