+++ src/generic/stage1/json_string_scanner.h
