+++ src/generic/stage1/json_structural_indexer.h
