+++ src/generic/stage2/streaming_structural_parser.h
