+++ src/generic/stage2/structural_iterator.h
