+++ src/generic/stage2/structural_parser.h
